+++ scripts/core/components/exports.js
@@ -3,5 +3,6 @@ export { default as Camera } from './camera.js';
 export { default as Clock }  from './clock.js';
 export { default as Frame }  from './frame.js';
 export { default as Hud }    from './hud.js';
+export { default as Layer }  from './layer.js';
 export { default as Room }   from './room.js';
 export { default as View }   from './view.js';